added rarray io, and definition files separating from implementation
